Extract adminapi and apiserver routes into internal/
