+++ cmd/apiserver/main.go
@@ -23,18 +23,11 @@ import (
 	"os"
 	"strconv"
 
-	"github.com/google/exposure-notifications-verification-server/pkg/api"
+	"github.com/google/exposure-notifications-verification-server/internal/routes"
 	"github.com/google/exposure-notifications-verification-server/pkg/buildinfo"
 	"github.com/google/exposure-notifications-verification-server/pkg/cache"
 	"github.com/google/exposure-notifications-verification-server/pkg/config"
-	"github.com/google/exposure-notifications-verification-server/pkg/controller"
-	"github.com/google/exposure-notifications-verification-server/pkg/controller/certapi"
-	"github.com/google/exposure-notifications-verification-server/pkg/controller/middleware"
-	"github.com/google/exposure-notifications-verification-server/pkg/controller/verifyapi"
-	"github.com/google/exposure-notifications-verification-server/pkg/database"
 	"github.com/google/exposure-notifications-verification-server/pkg/ratelimit"
-	"github.com/google/exposure-notifications-verification-server/pkg/ratelimit/limitware"
-	"github.com/google/exposure-notifications-verification-server/pkg/render"
 
 	"github.com/google/exposure-notifications-server/pkg/keys"
 	"github.com/google/exposure-notifications-server/pkg/logging"
@@ -42,8 +35,6 @@ import (
 	"github.com/google/exposure-notifications-server/pkg/server"
 
 	"github.com/gorilla/handlers"
-	"github.com/gorilla/mux"
-	"github.com/mikehelmick/go-chaff"
 	"github.com/sethvargo/go-signalcontext"
 )
 
@@ -85,7 +76,6 @@ func realMain(ctx context.Context) error {
 		return fmt.Errorf("error initializing observability exporter: %w", err)
 	}
 	defer oe.Close()
-	ctx, obs := middleware.WithObservability(ctx)
 	logger.Infow("observability exporter", "config", oeConfig)
 
 	// Setup cacher
@@ -105,6 +95,13 @@ func realMain(ctx context.Context) error {
 	}
 	defer db.Close()
 
+	// Setup rate limiter
+	limiterStore, err := ratelimit.RateLimiterFor(ctx, &cfg.RateLimit)
+	if err != nil {
+		return fmt.Errorf("failed to create limiter: %w", err)
+	}
+	defer limiterStore.Close(ctx)
+
 	// Setup signers
 	tokenSigner, err := keys.KeyManagerFor(ctx, &cfg.TokenSigning.Keys)
 	if err != nil {
@@ -115,120 +112,17 @@ func realMain(ctx context.Context) error {
 		return fmt.Errorf("failed to create certificate key manager: %w", err)
 	}
 
-	// Create the router
-	r := mux.NewRouter()
-
-	// Common observability context
-	r.Use(obs)
-
-	// Rate limiting
-	limiterStore, err := ratelimit.RateLimiterFor(ctx, &cfg.RateLimit)
+	// Setup routes
+	mux, err := routes.APIServer(ctx, cfg, db, cacher, limiterStore, tokenSigner, certificateSigner)
 	if err != nil {
-		return fmt.Errorf("failed to create limiter: %w", err)
-	}
-	defer limiterStore.Close(ctx)
-
-	// Note that rate limiting is installed _after_ the chaff middleware because
-	// we do not want chaff requests to count towards rate-limiting quota.
-	httplimiter, err := limitware.NewMiddleware(ctx, limiterStore,
-		limitware.APIKeyFunc(ctx, db, "apiserver:ratelimit:", cfg.RateLimit.HMACKey),
-		limitware.AllowOnError(false))
-	if err != nil {
-		return fmt.Errorf("failed to create limiter middleware: %w", err)
-	}
-	rateLimit := httplimiter.Handle
-
-	// Install common security headers
-	r.Use(middleware.SecureHeaders(cfg.DevMode, "json"))
-
-	// Enable debug headers
-	processDebug := middleware.ProcessDebug()
-	r.Use(processDebug)
-
-	// Create the renderer
-	h, err := render.New(ctx, "", cfg.DevMode)
-	if err != nil {
-		return fmt.Errorf("failed to create renderer: %w", err)
+		return fmt.Errorf("failed to setup routes: %w", err)
 	}
 
-	// Request ID injection
-	populateRequestID := middleware.PopulateRequestID(h)
-	r.Use(populateRequestID)
-
-	// Logger injection
-	populateLogger := middleware.PopulateLogger(logger)
-	r.Use(populateLogger)
-
-	// Other common middlewares
-	requireAPIKey := middleware.RequireAPIKey(cacher, db, h, []database.APIKeyType{
-		database.APIKeyTypeDevice,
-	})
-	processFirewall := middleware.ProcessFirewall(h, "apiserver")
-
-	r.Handle("/health", controller.HandleHealthz(ctx, &cfg.Database, h)).Methods("GET")
-
-	// Make verify chaff tracker.
-	verifyChaffTracker, err := chaff.NewTracker(chaff.NewJSONResponder(encodeVerifyResponse), chaff.DefaultCapacity)
-	if err != nil {
-		return fmt.Errorf("error creating verify chaffer: %v", err)
-	}
-	defer verifyChaffTracker.Close()
-
-	// Make cert chaff tracker.
-	certChaffTracker, err := chaff.NewTracker(chaff.NewJSONResponder(encodeCertificateResponse), chaff.DefaultCapacity)
-	if err != nil {
-		return fmt.Errorf("error creating cert chaffer: %v", err)
-	}
-	defer certChaffTracker.Close()
-
-	{
-		sub := r.PathPrefix("/api/verify").Subrouter()
-		sub.Use(requireAPIKey)
-		sub.Use(processFirewall)
-		sub.Use(middleware.ProcessChaff(db, verifyChaffTracker))
-		sub.Use(rateLimit)
-
-		// POST /api/verify
-		verifyapiController, err := verifyapi.New(ctx, cfg, db, h, tokenSigner)
-		if err != nil {
-			return fmt.Errorf("failed to create verify api controller: %w", err)
-		}
-		sub.Handle("", verifyapiController.HandleVerify()).Methods("POST")
-	}
-
-	{
-		sub := r.PathPrefix("/api/certificate").Subrouter()
-		sub.Use(requireAPIKey)
-		sub.Use(processFirewall)
-		sub.Use(middleware.ProcessChaff(db, certChaffTracker))
-		sub.Use(rateLimit)
-
-		// POST /api/certificate
-		certapiController, err := certapi.New(ctx, cfg, db, cacher, certificateSigner, h)
-		if err != nil {
-			return fmt.Errorf("failed to create certapi controller: %w", err)
-		}
-		sub.Handle("", certapiController.HandleCertificate()).Methods("POST")
-	}
-
+	// Run server
 	srv, err := server.New(cfg.Port)
 	if err != nil {
 		return fmt.Errorf("failed to create server: %w", err)
 	}
 	logger.Infow("server listening", "port", cfg.Port)
-	return srv.ServeHTTPHandler(ctx, handlers.CombinedLoggingHandler(os.Stdout, r))
-}
-
-// makePadFromChaff makes a Padding structure from chaff data.
-// Note, the random chaff data will be longer than necessary, so we shorten it.
-func makePadFromChaff(s string) api.Padding {
-	return api.Padding(s)
-}
-
-func encodeVerifyResponse(s string) interface{} {
-	return api.VerifyCodeResponse{Padding: makePadFromChaff(s)}
-}
-
-func encodeCertificateResponse(s string) interface{} {
-	return api.VerificationCertificateResponse{Padding: makePadFromChaff(s)}
+	return srv.ServeHTTPHandler(ctx, handlers.CombinedLoggingHandler(os.Stdout, mux))
 }
