+++ cmd/server/assets/users/show.html
@@ -62,8 +62,10 @@
             {{range $name, $permission := $permissions}}
               <li class="small">
                 {{if $userMembership.Can $permission}}
+                  <span class="sr-only">{{$user.Name}} can {{$name}}</span>
                   <span class="oi oi-circle-check text-success mr-1" aria-hidden="true"></span>
                 {{else}}
+                  <span class="sr-only">{{$user.Name}} cannot {{$name}}</span>
                   <span class="oi oi-circle-x text-muted mr-1" aria-hidden="true"></span>
                 {{end}}
                 {{$name}}
